requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.9.2
SQLAlchemy[asyncio]==2.0.43
aiosqlite==0.21.0
aiomysql==0.2.0
python-dotenv==1.0.1
numpy==2.3.0
pandas==2.3.2
httpx==0.28.1
jinja2==3.1.6
orjson==3.11.2
langgraph==0.6.6
python-multipart==0.0.6
